Make the allocator trait more similar to the standard one.
